refactor: data into data.model.ts and overview tab into new component
Prep for doing routes more nicelier
