+++ server/services/deployment/s-keyman-com.ts
@@ -5,7 +5,7 @@
 import httpget from "../../util/httpget";
 import DataService from "../data-service";
 
-const HOST='com-keyman-s.languagetechnology.org';
+const HOST='s.keyman.com';
 const PATH='/api/kmwversion.php';
 
 /*let debug_count = 275;
